+++ svc/pkg/user-identity/ops/delete/src/lib.rs
@@ -21,7 +21,9 @@ async fn handle(
 	)
 	.await?;
 
-	ctx.cache().purge("user_identity.emails", user_ids).await?;
+	ctx.cache()
+		.purge("user_identity.identities", user_ids)
+		.await?;
 
 	Ok(user_identity::delete::Response {})
 }