+++ CHANGELOG.md
@@ -52,6 +52,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   **Infra** NATS cluster not getting enabled
 -   **Infra** Redis Kubernetes error when using non-Kubernetes provider
 -   **api-helper** Remove excess logging
+-   `user_identity.identities` not getting purged on create & delete
 
 ## [23.2.0-rc.1] - 2023-12-01